Adding hash to prevent duplicate keys.
