+++ server.js
@@ -1,11 +1,10 @@
 const express = require('express');
-const redis = require('redis');
 const Redis = require('ioredis');
+const crypto = require('crypto'); // Added for creating hash of long URL
 
 // Create Express app
 const app = express();
-const port = 3000;
-const PORT = process.env.PORT || 3000;
+const port = process.env.PORT || 3000;
 
 // Create Redis client
 const client = new Redis({
@@ -13,117 +12,57 @@ const client = new Redis({
   port: 6379,
 });
 
-// In-memory store for URL mappings
-// const urlStore = {};
-
 // Middleware to parse JSON requests
 app.use(express.json());
 
-// Generate a random alphanumeric string as short code
-function generateShortCode() {
-  const chars = 'abcdefghijklmnopqrstuvwxyzABCDEFGHIJKLMNOPQRSTUVWXYZ0123456789';
-  let shortCode = '';
-  for (let i = 0; i < 6; i++) {
-    shortCode += chars.charAt(Math.floor(Math.random() * chars.length));
-  }
-  return shortCode;
+// Function to hash long URL to a fixed length short code
+function hashLongUrl(longUrl) {
+  // Using SHA-256 and slicing to get a shorter part of the hash
+  return crypto.createHash('sha256').update(longUrl).digest('hex').slice(0, 6);
 }
 
-// POST endpoint to save data to Redis
-app.post('/data', (req, res) => {
-  const { key, value } = req.body;
-
-  // Check if key and value are provided
-  if (!key || !value) {
-    return res.status(400).json({ error: 'Both key and value are required' });
+// Shorten URL endpoint
+app.post('/shorten', async (req, res) => {
+  const { longUrl } = req.body;
+  if (!longUrl) {
+    return res.status(400).json({ error: 'A long URL is required' });
   }
 
-  // Save data to Redis
-  client.set(key, value, (err, reply) => {
-    if (err) {
-      if (err instanceof redis.errors.ClientClosedError) {
-        // Handle ClientClosedError
-        return res.status(500).json({ error: 'Redis connection closed' });
-      } else {
-        console.error(err);
-        return res.status(500).json({ error: 'Internal Server Error' });
-      }
+  // Check if a short code for the long URL already exists
+  let shortCode = await client.get(`longUrl:${longUrl}`);
+  if (!shortCode) {
+    // If not, generate a new short code based on a hash of the long URL
+    shortCode = hashLongUrl(longUrl);
+    // Check if the generated shortCode already maps to a different long URL (collision)
+    const existingLongUrl = await client.get(`shortCode:${shortCode}`);
+    if (existingLongUrl && existingLongUrl !== longUrl) {
+      // Handle hash collision (very unlikely with a good hashing strategy and large enough hash space)
+      return res.status(500).json({ error: 'Short code generation collision. Try again.' });
     }
-    res.status(201).json({ message: 'Data saved successfully' });
-  });
-});
 
-// Shorten URL endpoint
-app.post('/shorten', (req, res) => {
-  const longUrl = req.body.longUrl;
-  const shortCode = generateShortCode();
+    // Save mappings from shortCode to longUrl and vice versa
+    await client.set(`shortCode:${shortCode}`, longUrl);
+    await client.set(`longUrl:${longUrl}`, shortCode);
+  }
+
   const shortUrl = `http://localhost:${port}/${shortCode}`;
-
-  // Save data to Redis
-  client.set(shortCode, longUrl, (err, reply) => {
-    if (err) {
-      console.error(err);
-      return res.status(500).json({ error: 'Internal Server Error' });
-    }
-    console.log(reply); // Reply from Redis
-
-    res.json({ shortUrl });
-  });
-});
-
-
-// GET endpoint to retrieve data from Redis
-app.get('/data/:key', (req, res) => {
-  const { key } = req.params;
-
-  // Retrieve data from Redis
-  client.get(key, (err, reply) => {
-    if (err) {
-      console.error(err);
-      return res.status(500).json({ error: 'Internal Server Error' });
-    }
-
-    if (!reply) {
-      return res.status(404).json({ error: 'Data not found' });
-    }
-
-    res.json({ key, value: reply });
-  });
+  res.json({ shortUrl });
 });
 
 // Redirect endpoint
-app.get('/:shortCode', (req, res) => {
-  const shortCode = req.params.shortCode;
-  client.get(shortCode, (err, longUrl) => {
-    if (err) {
-      console.error(err);
-      return res.status(500).json({ error: 'Internal Server Error' });
-    }
+app.get('/:shortCode', async (req, res) => {
+  const { shortCode } = req.params;
+
+  try {
+    const longUrl = await client.get(`shortCode:${shortCode}`);
     if (!longUrl) {
       return res.status(404).json({ error: 'URL not found' });
     }
     res.redirect(longUrl);
-  });
-});
-
-// DELETE endpoint to delete a mapping from Redis
-app.delete('/data/:key', (req, res) => {
-  const { key } = req.params;
-
-  // Delete data from Redis
-  client.del(key, (err, reply) => {
-    if (err) {
-      console.error(err);
-      return res.status(500).json({ error: 'Internal Server Error' });
-    }
-
-    if (reply === 0) {
-      // If reply is 0, it means the key does not exist in Redis
-      return res.status(404).json({ error: 'Data not found' });
-    }
-
-    res.status(204).end(); // No content, successful deletion
-  });
+  } catch (err) {
+    console.error(err);
+    res.status(500).json({ error: 'Internal Server Error' });
+  }
 });
 
 // Start the server
